requirements.txt
Adafruit-Blinka==8.39.1
adafruit-circuitpython-74hc595==1.4.4
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-charlcd==3.4.10
adafruit-circuitpython-connectionmanager==1.0.1
adafruit-circuitpython-mcp230xx==2.5.13
adafruit-circuitpython-requests==3.2.5
adafruit-circuitpython-typing==1.10.3
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
av==12.0.0
azure-core==1.30.1
azure-storage-blob==12.19.1
builtwith==1.3.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
filelock==3.13.4
fsspec==2024.3.1
gTTS==2.5.1
huggingface_hub==0.22.2
idna==3.7
isodate==0.6.1
Jinja2==3.1.3
kms==1.0.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
picamera==1.13
picamera2==0.3.18
pidng==4.0.9
piexif==1.1.3
pillow==10.3.0
pycparser==2.22
pyftdi==0.55.4
pygame==2.5.2
pyserial==3.5
python-prctl==1.8.1
pyusb==1.2.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
rpi-kms==0.1a1
rpi-libcamera==0.1a3
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
safetensors==0.4.3
simplejpeg==1.7.2
six==1.16.0
spidev==3.6
sympy==1.12
sysv-ipc==1.1.0
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
transformers==4.40.0
typing_extensions==4.11.0
urllib3==2.2.1
v4l2-python3==0.3.4